requirements.txt
pandas
tqdm
matplotlib
numpy
scikit-learn
dm-haiku
optax
hdbscan
